+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.2.pre - Unreleased
 
+- Added `IGMarkets::SprintMarketPosition#formatted_size`
 - Fixed `IGMarkets::RequestFailedError#message`
 
 ### 0.1 - April 8, 2016